Fix Player Punch Animations Pt1
Unfortunately, player 0 and player 1 punching animations are of a different (older?) animation than the rest of the players.  Unfortunately that is just what is available.  The sprites do not quite line up with the new animations naturally. And the torso rotation, being like no other, animation means I can't borrow from other sprites to make ones that are the same.   This is a first part to trying to get everything to align together.